refactor(ide): extract language features into separate modules
Extract the remaining language feature implementations from lib.rs into
dedicated modules for better organization and maintainability:

- goto_definition.rs - Symbol navigation (fields, types, fragments, etc.)
- references.rs - Find references for fragments, types, and fields
- symbols.rs - Document and workspace symbol providers
- code_lenses.rs - Fragment reference counts and deprecated field lenses
- semantic_tokens.rs - Syntax highlighting with deprecation awareness

This reduces lib.rs from 6361 to 5067 lines (~20% reduction) while
maintaining the same thin wrapper pattern established for hover and
completion modules.
