+++ crates/graphql-ide/src/lib.rs
@@ -51,8 +51,13 @@ pub(crate) mod symbol;
 mod types;
 
 // Feature modules
+mod code_lenses;
 mod completion;
+mod goto_definition;
 mod hover;
+mod references;
+mod semantic_tokens;
+mod symbols;
 
 // Re-export types from the types module
 pub use types::{
@@ -64,21 +69,12 @@ pub use types::{
 
 // Re-export helpers for internal use
 use helpers::{
-    adjust_range_for_line_offset, convert_diagnostic, find_argument_definition_in_tree,
-    find_block_for_position, find_field_name_at_offset, find_field_usages_in_parse,
-    find_fragment_definition_in_parse, find_fragment_spreads_in_parse,
-    find_operation_definition_in_tree, find_type_definition_in_parse,
-    find_type_references_in_parse, find_variable_definition_in_tree, format_type_ref,
-    offset_range_to_range, path_to_file_uri, position_to_offset,
+    adjust_range_for_line_offset, convert_diagnostic, offset_range_to_range, path_to_file_uri,
 };
 // Re-export for use in symbol module
 pub use helpers::unwrap_type_to_name;
 
-use symbol::{
-    extract_all_definitions, find_field_definition_full_range, find_fragment_definition_full_range,
-    find_operation_definition_ranges, find_parent_type_at_offset, find_schema_field_parent_type,
-    find_symbol_at_offset, find_type_definition_full_range, Symbol,
-};
+use symbol::{find_fragment_definition_full_range, find_operation_definition_ranges};
 
 // Re-export database types that IDE layer needs
 pub use graphql_base_db::FileKind;
@@ -1178,65 +1174,9 @@ impl Analysis {
     ///
     /// Returns tokens for syntax highlighting with semantic information,
     /// including deprecation status for fields.
-    #[allow(clippy::too_many_lines)]
     pub fn semantic_tokens(&self, file: &FilePath) -> Vec<SemanticToken> {
-        let (content, metadata) = {
-            let registry = self.registry.read();
-
-            // Look up FileId from FilePath
-            let Some(file_id) = registry.get_file_id(file) else {
-                return Vec::new();
-            };
-
-            // Get FileContent and FileMetadata
-            let Some(content) = registry.get_content(file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return Vec::new();
-            };
-            drop(registry);
-
-            (content, metadata)
-        };
-
-        // Parse the file
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-        if parse.has_errors() {
-            // Don't provide semantic tokens for files with parse errors
-            return Vec::new();
-        }
-
-        // Get schema types if available (for deprecation checks)
-        // schema_types returns a reference due to Salsa's returns(ref) optimization
-        let schema_types: Option<&std::collections::HashMap<Arc<str>, graphql_hir::TypeDef>> = self
-            .project_files
-            .map(|pf| graphql_hir::schema_types(&self.db, pf));
-
-        let mut tokens = Vec::new();
-
-        // Unified: collect tokens from all documents (works for both pure GraphQL and TS/JS)
-        #[allow(clippy::cast_possible_truncation)]
-        for doc in parse.documents() {
-            let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-            collect_semantic_tokens_from_document(
-                &doc.tree.document(),
-                &doc_line_index,
-                doc.line_offset as u32,
-                schema_types,
-                &mut tokens,
-            );
-        }
-
-        // Sort tokens by position (required for delta encoding)
-        tokens.sort_by(|a, b| {
-            a.start
-                .line
-                .cmp(&b.start.line)
-                .then_with(|| a.start.character.cmp(&b.start.character))
-        });
-
-        tokens
+        let registry = self.registry.read();
+        semantic_tokens::semantic_tokens(&self.db, &registry, self.project_files, file)
     }
 
     /// Get project-wide lint diagnostics (e.g., unused fields, unique names)
@@ -1555,261 +1495,9 @@ impl Analysis {
     /// Get goto definition locations for the symbol at a position
     ///
     /// Returns the definition location(s) for types, fields, fragments, etc.
-    #[allow(clippy::too_many_lines)]
     pub fn goto_definition(&self, file: &FilePath, position: Position) -> Option<Vec<Location>> {
-        let (content, metadata) = {
-            let registry = self.registry.read();
-
-            let file_id = registry.get_file_id(file)?;
-
-            let content = registry.get_content(file_id)?;
-            let metadata = registry.get_metadata(file_id)?;
-            drop(registry);
-
-            (content, metadata)
-        };
-
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        let (block_context, adjusted_position) = find_block_for_position(&parse, position)?;
-
-        tracing::debug!(
-            "Goto definition: original position {:?}, block line_offset {}, adjusted position {:?}",
-            position,
-            block_context.line_offset,
-            adjusted_position
-        );
-
-        // Create line index from block source (all documents now have source)
-        let block_line_index = graphql_syntax::LineIndex::new(block_context.block_source);
-        let offset = position_to_offset(&block_line_index, adjusted_position)?;
-
-        let symbol = find_symbol_at_offset(block_context.tree, offset)?;
-
-        let project_files = self.project_files?;
-
-        match symbol {
-            Symbol::FieldName { name } => {
-                let schema_types = graphql_hir::schema_types(&self.db, project_files);
-
-                // Try to find parent type from executable document (operation/fragment)
-                // or fall back to schema definition (type/interface)
-                let parent_type_name = if let Some(parent_ctx) =
-                    find_parent_type_at_offset(block_context.tree, offset)
-                {
-                    // Use walk_type_stack_to_offset to properly resolve the parent type,
-                    // which handles inline fragments correctly
-                    symbol::walk_type_stack_to_offset(
-                        block_context.tree,
-                        schema_types,
-                        offset,
-                        &parent_ctx.root_type,
-                    )?
-                } else {
-                    // We're on a schema field definition - get the parent type name
-                    symbol::find_schema_field_parent_type(block_context.tree, offset)?
-                };
-
-                tracing::debug!(
-                    "Field '{}' - resolved parent type '{}'",
-                    name,
-                    parent_type_name
-                );
-
-                schema_types.get(parent_type_name.as_str())?;
-
-                let registry = self.registry.read();
-                let schema_file_ids = project_files.schema_file_ids(&self.db).ids(&self.db);
-
-                for file_id in schema_file_ids.iter() {
-                    let Some(schema_content) = registry.get_content(*file_id) else {
-                        continue;
-                    };
-                    let Some(schema_metadata) = registry.get_metadata(*file_id) else {
-                        continue;
-                    };
-                    let Some(file_path) = registry.get_path(*file_id) else {
-                        continue;
-                    };
-
-                    let schema_parse =
-                        graphql_syntax::parse(&self.db, schema_content, schema_metadata);
-
-                    for doc in schema_parse.documents() {
-                        if let Some(ranges) =
-                            find_field_definition_full_range(doc.tree, &parent_type_name, &name)
-                        {
-                            let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                            let range = offset_range_to_range(
-                                &doc_line_index,
-                                ranges.name_start,
-                                ranges.name_end,
-                            );
-                            #[allow(clippy::cast_possible_truncation)]
-                            let adjusted_range =
-                                adjust_range_for_line_offset(range, doc.line_offset as u32);
-                            return Some(vec![Location::new(file_path, adjusted_range)]);
-                        }
-                    }
-                }
-
-                // Field definition not found
-                None
-            }
-            Symbol::FragmentSpread { name } => {
-                let fragments = graphql_hir::all_fragments(&self.db, project_files);
-
-                tracing::debug!(
-                    "Looking for fragment '{}', available fragments: {:?}",
-                    name,
-                    fragments.keys().collect::<Vec<_>>()
-                );
-
-                let fragment = fragments.get(name.as_str())?;
-
-                let registry = self.registry.read();
-
-                tracing::debug!(
-                    "Looking up path for fragment '{}' with FileId {:?}",
-                    name,
-                    fragment.file_id
-                );
-                let all_ids = registry.all_file_ids();
-                tracing::debug!("Registry has {} files", all_ids.len());
-                tracing::debug!("Registry FileIds: {:?}", all_ids);
-
-                let Some(file_path) = registry.get_path(fragment.file_id) else {
-                    tracing::error!(
-                        "FileId {:?} not found in registry for fragment '{}'",
-                        fragment.file_id,
-                        name
-                    );
-                    return None;
-                };
-                let def_content = registry.get_content(fragment.file_id)?;
-                let def_metadata = registry.get_metadata(fragment.file_id)?;
-                drop(registry);
-
-                let def_parse = graphql_syntax::parse(&self.db, def_content, def_metadata);
-
-                let range =
-                    find_fragment_definition_in_parse(&def_parse, &name, def_content, &self.db)?;
-
-                Some(vec![Location::new(file_path, range)])
-            }
-            Symbol::TypeName { name } => {
-                let schema_ids = project_files.schema_file_ids(&self.db).ids(&self.db);
-                let registry = self.registry.read();
-
-                for file_id in schema_ids.iter() {
-                    let Some(schema_content) = registry.get_content(*file_id) else {
-                        continue;
-                    };
-                    let Some(schema_metadata) = registry.get_metadata(*file_id) else {
-                        continue;
-                    };
-                    let Some(file_path) = registry.get_path(*file_id) else {
-                        continue;
-                    };
-
-                    let schema_parse =
-                        graphql_syntax::parse(&self.db, schema_content, schema_metadata);
-
-                    if let Some(range) = find_type_definition_in_parse(
-                        &schema_parse,
-                        &name,
-                        schema_content,
-                        &self.db,
-                    ) {
-                        return Some(vec![Location::new(file_path, range)]);
-                    }
-                }
-
-                // Type definition not found
-                None
-            }
-            Symbol::VariableReference { name } => {
-                let block_line_index = graphql_syntax::LineIndex::new(block_context.block_source);
-                let range = find_variable_definition_in_tree(
-                    block_context.tree,
-                    &name,
-                    &block_line_index,
-                    block_context.line_offset,
-                );
-
-                if let Some(range) = range {
-                    let registry = self.registry.read();
-                    let file_id = registry.get_file_id(file)?;
-                    let file_path = registry.get_path(file_id)?;
-                    return Some(vec![Location::new(file_path, range)]);
-                }
-                None
-            }
-            Symbol::ArgumentName { name } => {
-                let parent_context = find_parent_type_at_offset(block_context.tree, offset)?;
-                let schema_types = graphql_hir::schema_types(&self.db, project_files);
-
-                let field_name = find_field_name_at_offset(block_context.tree, offset)?;
-
-                let parent_type_name = symbol::walk_type_stack_to_offset(
-                    block_context.tree,
-                    schema_types,
-                    offset,
-                    &parent_context.root_type,
-                )?;
-
-                let registry = self.registry.read();
-                let schema_file_ids = project_files.schema_file_ids(&self.db).ids(&self.db);
-
-                for file_id in schema_file_ids.iter() {
-                    let Some(schema_content) = registry.get_content(*file_id) else {
-                        continue;
-                    };
-                    let Some(schema_metadata) = registry.get_metadata(*file_id) else {
-                        continue;
-                    };
-                    let Some(file_path) = registry.get_path(*file_id) else {
-                        continue;
-                    };
-
-                    let schema_parse =
-                        graphql_syntax::parse(&self.db, schema_content, schema_metadata);
-
-                    for doc in schema_parse.documents() {
-                        let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                        #[allow(clippy::cast_possible_truncation)]
-                        if let Some(range) = find_argument_definition_in_tree(
-                            doc.tree,
-                            &parent_type_name,
-                            &field_name,
-                            &name,
-                            &doc_line_index,
-                            doc.line_offset as u32,
-                        ) {
-                            return Some(vec![Location::new(file_path, range)]);
-                        }
-                    }
-                }
-                None
-            }
-            Symbol::OperationName { name } => {
-                let block_line_index = graphql_syntax::LineIndex::new(block_context.block_source);
-                let range = find_operation_definition_in_tree(
-                    block_context.tree,
-                    &name,
-                    &block_line_index,
-                    block_context.line_offset,
-                );
-
-                if let Some(range) = range {
-                    let registry = self.registry.read();
-                    let file_id = registry.get_file_id(file)?;
-                    let file_path = registry.get_path(file_id)?;
-                    return Some(vec![Location::new(file_path, range)]);
-                }
-                None
-            }
-        }
+        let registry = self.registry.read();
+        goto_definition::goto_definition(&self.db, &registry, self.project_files, file, position)
     }
 
     /// Find all references to the symbol at a position
@@ -1821,48 +1509,15 @@ impl Analysis {
         position: Position,
         include_declaration: bool,
     ) -> Option<Vec<Location>> {
-        let (content, metadata) = {
-            let registry = self.registry.read();
-
-            let file_id = registry.get_file_id(file)?;
-
-            let content = registry.get_content(file_id)?;
-            let metadata = registry.get_metadata(file_id)?;
-            drop(registry);
-
-            (content, metadata)
-        };
-
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        let (block_context, adjusted_position) = find_block_for_position(&parse, position)?;
-
-        tracing::debug!(
-            "Find references: original position {:?}, block line_offset {}, adjusted position {:?}",
+        let registry = self.registry.read();
+        references::find_references(
+            &self.db,
+            &registry,
+            self.project_files,
+            file,
             position,
-            block_context.line_offset,
-            adjusted_position
-        );
-
-        // Create line index from block source (all documents now have source)
-        let block_line_index = graphql_syntax::LineIndex::new(block_context.block_source);
-        let offset = position_to_offset(&block_line_index, adjusted_position)?;
-
-        let symbol = find_symbol_at_offset(block_context.tree, offset)?;
-
-        match symbol {
-            Symbol::FragmentSpread { name } => {
-                Some(self.find_fragment_references(&name, include_declaration))
-            }
-            Symbol::TypeName { name } => {
-                Some(self.find_type_references(&name, include_declaration))
-            }
-            Symbol::FieldName { name } => {
-                let parent_type = find_schema_field_parent_type(block_context.tree, offset)?;
-                Some(self.find_field_references(&parent_type, &name, include_declaration))
-            }
-            _ => None,
-        }
+            include_declaration,
+        )
     }
 
     /// Find all references to a fragment
@@ -1871,221 +1526,14 @@ impl Analysis {
         fragment_name: &str,
         include_declaration: bool,
     ) -> Vec<Location> {
-        let mut locations = Vec::new();
-
-        let Some(project_files) = self.project_files else {
-            return locations;
-        };
-
-        let fragments = graphql_hir::all_fragments(&self.db, project_files);
-
-        if include_declaration {
-            if let Some(fragment) = fragments.get(fragment_name) {
-                let registry = self.registry.read();
-                let file_path = registry.get_path(fragment.file_id);
-                let def_content = registry.get_content(fragment.file_id);
-                let def_metadata = registry.get_metadata(fragment.file_id);
-                drop(registry);
-
-                if let (Some(file_path), Some(def_content), Some(def_metadata)) =
-                    (file_path, def_content, def_metadata)
-                {
-                    let def_parse = graphql_syntax::parse(&self.db, def_content, def_metadata);
-
-                    if let Some(range) = find_fragment_definition_in_parse(
-                        &def_parse,
-                        fragment_name,
-                        def_content,
-                        &self.db,
-                    ) {
-                        locations.push(Location::new(file_path, range));
-                    }
-                }
-            }
-        }
-
-        // Search through all document files for fragment spreads
-        let doc_ids = project_files.document_file_ids(&self.db).ids(&self.db);
-
-        for file_id in doc_ids.iter() {
-            let Some((content, metadata)) =
-                graphql_base_db::file_lookup(&self.db, project_files, *file_id)
-            else {
-                continue;
-            };
-
-            let registry = self.registry.read();
-            let file_path = registry.get_path(*file_id);
-            drop(registry);
-
-            let Some(file_path) = file_path else {
-                continue;
-            };
-
-            let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-            let spread_ranges =
-                find_fragment_spreads_in_parse(&parse, fragment_name, content, &self.db);
-
-            for range in spread_ranges {
-                locations.push(Location::new(file_path.clone(), range));
-            }
-        }
-
-        locations
-    }
-
-    /// Find all references to a type
-    fn find_type_references(&self, type_name: &str, include_declaration: bool) -> Vec<Location> {
-        let mut locations = Vec::new();
-
-        let Some(project_files) = self.project_files else {
-            return locations;
-        };
-
-        let types = graphql_hir::schema_types(&self.db, project_files);
-
-        if include_declaration {
-            if let Some(type_def) = types.get(type_name) {
-                let registry = self.registry.read();
-                let file_path = registry.get_path(type_def.file_id);
-                let def_content = registry.get_content(type_def.file_id);
-                let def_metadata = registry.get_metadata(type_def.file_id);
-                drop(registry);
-
-                if let (Some(file_path), Some(def_content), Some(def_metadata)) =
-                    (file_path, def_content, def_metadata)
-                {
-                    let def_parse = graphql_syntax::parse(&self.db, def_content, def_metadata);
-
-                    if let Some(range) =
-                        find_type_definition_in_parse(&def_parse, type_name, def_content, &self.db)
-                    {
-                        locations.push(Location::new(file_path, range));
-                    }
-                }
-            }
-        }
-
-        let schema_ids = project_files.schema_file_ids(&self.db).ids(&self.db);
-
-        for file_id in schema_ids.iter() {
-            let Some((content, metadata)) =
-                graphql_base_db::file_lookup(&self.db, project_files, *file_id)
-            else {
-                continue;
-            };
-
-            let registry = self.registry.read();
-            let file_path = registry.get_path(*file_id);
-            drop(registry);
-
-            let Some(file_path) = file_path else {
-                continue;
-            };
-
-            let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-            let type_ranges = find_type_references_in_parse(&parse, type_name, content, &self.db);
-
-            for range in type_ranges {
-                locations.push(Location::new(file_path.clone(), range));
-            }
-        }
-
-        locations
-    }
-
-    /// Find all references to a field on a specific type
-    fn find_field_references(
-        &self,
-        type_name: &str,
-        field_name: &str,
-        include_declaration: bool,
-    ) -> Vec<Location> {
-        let mut locations = Vec::new();
-
-        let Some(project_files) = self.project_files else {
-            return locations;
-        };
-
-        let schema_types = graphql_hir::schema_types(&self.db, project_files);
-
-        if include_declaration {
-            let schema_ids = project_files.schema_file_ids(&self.db).ids(&self.db);
-
-            for file_id in schema_ids.iter() {
-                let Some((content, metadata)) =
-                    graphql_base_db::file_lookup(&self.db, project_files, *file_id)
-                else {
-                    continue;
-                };
-
-                let registry = self.registry.read();
-                let file_path = registry.get_path(*file_id);
-                drop(registry);
-
-                let Some(file_path) = file_path else {
-                    continue;
-                };
-
-                let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-                'schema_search: for doc in parse.documents() {
-                    if let Some(ranges) =
-                        find_field_definition_full_range(doc.tree, type_name, field_name)
-                    {
-                        let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                        let range = offset_range_to_range(
-                            &doc_line_index,
-                            ranges.name_start,
-                            ranges.name_end,
-                        );
-                        #[allow(clippy::cast_possible_truncation)]
-                        let adjusted_range =
-                            adjust_range_for_line_offset(range, doc.line_offset as u32);
-                        locations.push(Location::new(file_path, adjusted_range));
-                        break 'schema_search; // Field definition found
-                    }
-                }
-            }
-        }
-
-        // Search through all document files for field usages
-        let doc_ids = project_files.document_file_ids(&self.db).ids(&self.db);
-
-        for file_id in doc_ids.iter() {
-            let Some((content, metadata)) =
-                graphql_base_db::file_lookup(&self.db, project_files, *file_id)
-            else {
-                continue;
-            };
-
-            let registry = self.registry.read();
-            let file_path = registry.get_path(*file_id);
-            drop(registry);
-
-            let Some(file_path) = file_path else {
-                continue;
-            };
-
-            let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-            let field_ranges = find_field_usages_in_parse(
-                &parse,
-                type_name,
-                field_name,
-                schema_types,
-                content,
-                &self.db,
-            );
-
-            for range in field_ranges {
-                locations.push(Location::new(file_path.clone(), range));
-            }
-        }
-
-        locations
+        let registry = self.registry.read();
+        references::find_fragment_references(
+            &self.db,
+            &registry,
+            self.project_files,
+            fragment_name,
+            include_declaration,
+        )
     }
 
     /// Get code lenses for deprecated fields in a schema file
@@ -2093,199 +1541,17 @@ impl Analysis {
     /// Returns code lens information for each deprecated field definition,
     /// including the usage count and locations for navigation.
     pub fn deprecated_field_code_lenses(&self, file: &FilePath) -> Vec<CodeLensInfo> {
-        let mut code_lenses = Vec::new();
-
-        let Some(project_files) = self.project_files else {
-            return code_lenses;
-        };
-
-        // Get the file_id for this file
-        let file_id = {
-            let registry = self.registry.read();
-            registry.get_file_id(file)
-        };
-
-        let Some(file_id) = file_id else {
-            return code_lenses;
-        };
-
-        // Get schema types to find deprecated fields
-        let schema_types = graphql_hir::schema_types(&self.db, project_files);
-
-        // Get file content and metadata for line index
-        let (content, metadata) = {
-            let registry = self.registry.read();
-            let content = registry.get_content(file_id);
-            let metadata = registry.get_metadata(file_id);
-            (content, metadata)
-        };
-
-        let (Some(content), Some(_metadata)) = (content, metadata) else {
-            return code_lenses;
-        };
-
-        let line_index = graphql_syntax::line_index(&self.db, content);
-
-        // Iterate through all types and find deprecated fields in this file
-        for type_def in schema_types.values() {
-            // Only process types defined in the current file
-            if type_def.file_id != file_id {
-                continue;
-            }
-
-            for field in &type_def.fields {
-                if !field.is_deprecated {
-                    continue;
-                }
-
-                // Find usages of this deprecated field
-                let usage_locations = self.find_field_references(
-                    type_def.name.as_ref(),
-                    field.name.as_ref(),
-                    false, // don't include declaration
-                );
-
-                // Convert the field's name_range to editor coordinates
-                let name_start = field.name_range.start().into();
-                let name_end = field.name_range.end().into();
-                let range = offset_range_to_range(&line_index, name_start, name_end);
-
-                let mut code_lens = CodeLensInfo::new(
-                    range,
-                    type_def.name.as_ref(),
-                    field.name.as_ref(),
-                    usage_locations.len(),
-                    usage_locations,
-                );
-
-                if let Some(ref reason) = field.deprecation_reason {
-                    code_lens = code_lens.with_deprecation_reason(reason.as_ref());
-                }
-
-                code_lenses.push(code_lens);
-            }
-
-            // Note: Enum values with @deprecated could also be supported here
-            // by adding a find_enum_value_references method. Left for future work.
-        }
-
-        code_lenses
+        let registry = self.registry.read();
+        code_lenses::deprecated_field_code_lenses(&self.db, &registry, self.project_files, file)
     }
 
     /// Get document symbols for a file (hierarchical outline)
     ///
     /// Returns types, operations, and fragments with their fields as children.
     /// This powers the "Go to Symbol in Editor" (Cmd+Shift+O) feature.
-    #[allow(clippy::too_many_lines)]
     pub fn document_symbols(&self, file: &FilePath) -> Vec<DocumentSymbol> {
-        let (content, metadata, file_id) = {
-            let registry = self.registry.read();
-
-            let Some(file_id) = registry.get_file_id(file) else {
-                return Vec::new();
-            };
-
-            let Some(content) = registry.get_content(file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return Vec::new();
-            };
-            drop(registry);
-
-            (content, metadata, file_id)
-        };
-
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        let structure = graphql_hir::file_structure(&self.db, file_id, content, metadata);
-
-        let mut symbols = Vec::new();
-
-        for doc in parse.documents() {
-            let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-            #[allow(clippy::cast_possible_truncation)]
-            let doc_line_offset = doc.line_offset as u32;
-
-            let definitions = extract_all_definitions(doc.tree);
-
-            for (name, kind, ranges) in definitions {
-                let range = adjust_range_for_line_offset(
-                    offset_range_to_range(&doc_line_index, ranges.def_start, ranges.def_end),
-                    doc_line_offset,
-                );
-                let selection_range = adjust_range_for_line_offset(
-                    offset_range_to_range(&doc_line_index, ranges.name_start, ranges.name_end),
-                    doc_line_offset,
-                );
-
-                let symbol = match kind {
-                    "object" => {
-                        let children = get_field_children(
-                            &structure,
-                            &name,
-                            doc.tree,
-                            &doc_line_index,
-                            doc_line_offset,
-                        );
-                        DocumentSymbol::new(name, SymbolKind::Type, range, selection_range)
-                            .with_children(children)
-                    }
-                    "interface" => {
-                        let children = get_field_children(
-                            &structure,
-                            &name,
-                            doc.tree,
-                            &doc_line_index,
-                            doc_line_offset,
-                        );
-                        DocumentSymbol::new(name, SymbolKind::Interface, range, selection_range)
-                            .with_children(children)
-                    }
-                    "input" => {
-                        let children = get_field_children(
-                            &structure,
-                            &name,
-                            doc.tree,
-                            &doc_line_index,
-                            doc_line_offset,
-                        );
-                        DocumentSymbol::new(name, SymbolKind::Input, range, selection_range)
-                            .with_children(children)
-                    }
-                    "union" => DocumentSymbol::new(name, SymbolKind::Union, range, selection_range),
-                    "enum" => DocumentSymbol::new(name, SymbolKind::Enum, range, selection_range),
-                    "scalar" => {
-                        DocumentSymbol::new(name, SymbolKind::Scalar, range, selection_range)
-                    }
-                    "query" => DocumentSymbol::new(name, SymbolKind::Query, range, selection_range),
-                    "mutation" => {
-                        DocumentSymbol::new(name, SymbolKind::Mutation, range, selection_range)
-                    }
-                    "subscription" => {
-                        DocumentSymbol::new(name, SymbolKind::Subscription, range, selection_range)
-                    }
-                    "fragment" => {
-                        let detail = structure
-                            .fragments
-                            .iter()
-                            .find(|f| f.name.as_ref() == name)
-                            .map(|f| format!("on {}", f.type_condition));
-                        let mut sym =
-                            DocumentSymbol::new(name, SymbolKind::Fragment, range, selection_range);
-                        if let Some(d) = detail {
-                            sym = sym.with_detail(d);
-                        }
-                        sym
-                    }
-                    _ => continue,
-                };
-
-                symbols.push(symbol);
-            }
-        }
-
-        symbols
+        let registry = self.registry.read();
+        symbols::document_symbols(&self.db, &registry, file)
     }
 
     /// Search for workspace symbols matching a query
@@ -2293,78 +1559,8 @@ impl Analysis {
     /// Returns matching types, operations, and fragments across all files.
     /// This powers the "Go to Symbol in Workspace" (Cmd+T) feature.
     pub fn workspace_symbols(&self, query: &str) -> Vec<WorkspaceSymbol> {
-        let Some(project_files) = self.project_files else {
-            return Vec::new();
-        };
-
-        let query_lower = query.to_lowercase();
-        let mut symbols = Vec::new();
-
-        // Search types
-        let types = graphql_hir::schema_types(&self.db, project_files);
-        for (name, type_def) in types {
-            if name.to_lowercase().contains(&query_lower) {
-                if let Some(location) = self.get_type_location(type_def) {
-                    #[allow(clippy::match_same_arms)]
-                    let kind = match type_def.kind {
-                        graphql_hir::TypeDefKind::Object => SymbolKind::Type,
-                        graphql_hir::TypeDefKind::Interface => SymbolKind::Interface,
-                        graphql_hir::TypeDefKind::Union => SymbolKind::Union,
-                        graphql_hir::TypeDefKind::Enum => SymbolKind::Enum,
-                        graphql_hir::TypeDefKind::Scalar => SymbolKind::Scalar,
-                        graphql_hir::TypeDefKind::InputObject => SymbolKind::Input,
-                        _ => SymbolKind::Type, // fallback for future type kinds
-                    };
-
-                    symbols.push(WorkspaceSymbol::new(name.to_string(), kind, location));
-                }
-            }
-        }
-
-        // Search fragments
-        let fragments = graphql_hir::all_fragments(&self.db, project_files);
-        for (name, fragment) in fragments {
-            if name.to_lowercase().contains(&query_lower) {
-                if let Some(location) = self.get_fragment_location(fragment) {
-                    symbols.push(
-                        WorkspaceSymbol::new(name.to_string(), SymbolKind::Fragment, location)
-                            .with_container(format!("on {}", fragment.type_condition)),
-                    );
-                }
-            }
-        }
-
-        // Search operations from document files
-        let doc_ids = project_files.document_file_ids(&self.db).ids(&self.db);
-        for file_id in doc_ids.iter() {
-            let Some((content, metadata)) =
-                graphql_base_db::file_lookup(&self.db, project_files, *file_id)
-            else {
-                continue;
-            };
-            let structure = graphql_hir::file_structure(&self.db, *file_id, content, metadata);
-            for operation in structure.operations.iter() {
-                if let Some(op_name) = &operation.name {
-                    if op_name.to_lowercase().contains(&query_lower) {
-                        if let Some(location) = self.get_operation_location(operation) {
-                            #[allow(clippy::match_same_arms)]
-                            let kind = match operation.operation_type {
-                                graphql_hir::OperationType::Query => SymbolKind::Query,
-                                graphql_hir::OperationType::Mutation => SymbolKind::Mutation,
-                                graphql_hir::OperationType::Subscription => {
-                                    SymbolKind::Subscription
-                                }
-                                _ => SymbolKind::Query, // fallback for future operation types
-                            };
-
-                            symbols.push(WorkspaceSymbol::new(op_name.to_string(), kind, location));
-                        }
-                    }
-                }
-            }
-        }
-
-        symbols
+        let registry = self.registry.read();
+        symbols::workspace_symbols(&self.db, &registry, self.project_files, query)
     }
 
     /// Get schema statistics
@@ -2425,86 +1621,6 @@ impl Analysis {
         }
 
         stats
-    }
-
-    /// Get location for a type definition
-    fn get_type_location(&self, type_def: &graphql_hir::TypeDef) -> Option<Location> {
-        let registry = self.registry.read();
-        let file_path = registry.get_path(type_def.file_id)?;
-        let content = registry.get_content(type_def.file_id)?;
-        let metadata = registry.get_metadata(type_def.file_id)?;
-        drop(registry);
-
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        for doc in parse.documents() {
-            if let Some(ranges) = find_type_definition_full_range(doc.tree, &type_def.name) {
-                let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                #[allow(clippy::cast_possible_truncation)]
-                let range = adjust_range_for_line_offset(
-                    offset_range_to_range(&doc_line_index, ranges.name_start, ranges.name_end),
-                    doc.line_offset as u32,
-                );
-                return Some(Location::new(file_path, range));
-            }
-        }
-
-        None
-    }
-
-    /// Get location for a fragment definition
-    fn get_fragment_location(&self, fragment: &graphql_hir::FragmentStructure) -> Option<Location> {
-        let registry = self.registry.read();
-        let file_path = registry.get_path(fragment.file_id)?;
-        let content = registry.get_content(fragment.file_id)?;
-        let metadata = registry.get_metadata(fragment.file_id)?;
-        drop(registry);
-
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        for doc in parse.documents() {
-            if let Some(ranges) = find_fragment_definition_full_range(doc.tree, &fragment.name) {
-                let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                #[allow(clippy::cast_possible_truncation)]
-                let range = adjust_range_for_line_offset(
-                    offset_range_to_range(&doc_line_index, ranges.name_start, ranges.name_end),
-                    doc.line_offset as u32,
-                );
-                return Some(Location::new(file_path, range));
-            }
-        }
-
-        None
-    }
-
-    /// Get location for an operation definition
-    fn get_operation_location(
-        &self,
-        operation: &graphql_hir::OperationStructure,
-    ) -> Option<Location> {
-        let op_name = operation.name.as_ref()?;
-
-        let registry = self.registry.read();
-        let file_path = registry.get_path(operation.file_id)?;
-        let content = registry.get_content(operation.file_id)?;
-        let metadata = registry.get_metadata(operation.file_id)?;
-        drop(registry);
-
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        for doc in parse.documents() {
-            if let Some(ranges) = find_operation_definition_ranges(doc.tree, op_name) {
-                let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                #[allow(clippy::cast_possible_truncation)]
-                let range = adjust_range_for_line_offset(
-                    offset_range_to_range(&doc_line_index, ranges.name_start, ranges.name_end),
-                    doc.line_offset as u32,
-                );
-                return Some(Location::new(file_path, range));
-            }
-        }
-
-        None
     }
 
     /// Get fragment usage analysis for the project
@@ -2617,429 +1733,19 @@ impl Analysis {
     ///
     /// Returns code lenses for fragment definitions showing reference counts.
     pub fn code_lenses(&self, file: &FilePath) -> Vec<CodeLens> {
-        let (content, metadata, file_id) = {
-            let registry = self.registry.read();
-
-            let Some(file_id) = registry.get_file_id(file) else {
-                return Vec::new();
-            };
-
-            let Some(content) = registry.get_content(file_id) else {
-                return Vec::new();
-            };
-            let Some(metadata) = registry.get_metadata(file_id) else {
-                return Vec::new();
-            };
-            drop(registry);
-
-            (content, metadata, file_id)
-        };
-
-        // Verify project files are loaded (fragment_usages needs them)
-        if self.project_files.is_none() {
-            return Vec::new();
-        }
-
-        // Get fragment usage information
         let fragment_usages = self.fragment_usages();
-
-        // Get fragments defined in this file
-        let structure = graphql_hir::file_structure(&self.db, file_id, content, metadata);
-
-        let mut lenses = Vec::new();
-        let parse = graphql_syntax::parse(&self.db, content, metadata);
-
-        for fragment in structure.fragments.iter() {
-            // Find usage info for this fragment
-            let usage_count = fragment_usages
-                .iter()
-                .find(|u| u.name == fragment.name.as_ref())
-                .map_or(0, FragmentUsage::usage_count);
-
-            // Get the range for the fragment definition line
-            for doc in parse.documents() {
-                if let Some(ranges) = find_fragment_definition_full_range(doc.tree, &fragment.name)
-                {
-                    let doc_line_index = graphql_syntax::LineIndex::new(doc.source);
-                    #[allow(clippy::cast_possible_truncation)]
-                    let range = adjust_range_for_line_offset(
-                        offset_range_to_range(&doc_line_index, ranges.def_start, ranges.def_start),
-                        doc.line_offset as u32,
-                    );
-
-                    let title = if usage_count == 1 {
-                        "1 reference".to_string()
-                    } else {
-                        format!("{usage_count} references")
-                    };
-
-                    // Create command to show references when clicked
-                    let command = CodeLensCommand::new("editor.action.showReferences", &title)
-                        .with_arguments(vec![
-                            file.as_str().to_string(),
-                            format!("{}:{}", range.start.line, range.start.character),
-                            fragment.name.to_string(),
-                        ]);
-
-                    lenses.push(CodeLens::new(range, title).with_command(command));
-                    break; // Found the fragment, no need to check other documents
-                }
-            }
-        }
-
-        tracing::debug!(lens_count = lenses.len(), "code_lenses: returning");
-        lenses
+        let registry = self.registry.read();
+        code_lenses::code_lenses(
+            &self.db,
+            &registry,
+            self.project_files,
+            file,
+            &fragment_usages,
+        )
     }
 }
 
-// Helper functions are now in helpers.rs module
-
-/// Collect semantic tokens from a GraphQL document
-///
-/// Walks the document and emits tokens for fields, types, fragments, etc.
-/// Checks the schema to determine if fields are deprecated.
-#[allow(clippy::too_many_lines)]
-fn collect_semantic_tokens_from_document(
-    doc_cst: &apollo_parser::cst::Document,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-    schema_types: Option<&std::collections::HashMap<Arc<str>, graphql_hir::TypeDef>>,
-    tokens: &mut Vec<SemanticToken>,
-) {
-    use apollo_parser::cst::{self, CstNode};
-
-    for definition in doc_cst.definitions() {
-        match definition {
-            cst::Definition::OperationDefinition(operation) => {
-                // Emit token for operation keyword (query, mutation, subscription)
-                if let Some(op_type) = operation.operation_type() {
-                    if let Some(token) = op_type
-                        .query_token()
-                        .or_else(|| op_type.mutation_token())
-                        .or_else(|| op_type.subscription_token())
-                    {
-                        emit_token_for_syntax_token(
-                            &token,
-                            line_index,
-                            line_offset,
-                            SemanticTokenType::Keyword,
-                            SemanticTokenModifiers::NONE,
-                            tokens,
-                        );
-                    }
-                }
-
-                // Determine root type for field deprecation checks
-                let root_type_name = operation.operation_type().map_or("Query", |op_type| {
-                    if op_type.query_token().is_some() {
-                        "Query"
-                    } else if op_type.mutation_token().is_some() {
-                        "Mutation"
-                    } else if op_type.subscription_token().is_some() {
-                        "Subscription"
-                    } else {
-                        "Query"
-                    }
-                });
-
-                // Process selection set
-                if let Some(selection_set) = operation.selection_set() {
-                    collect_tokens_from_selection_set(
-                        &selection_set,
-                        Some(root_type_name),
-                        schema_types,
-                        line_index,
-                        line_offset,
-                        tokens,
-                    );
-                }
-            }
-            cst::Definition::FragmentDefinition(fragment) => {
-                // Emit token for "fragment" keyword
-                if let Some(fragment_token) = fragment.fragment_token() {
-                    emit_token_for_syntax_token(
-                        &fragment_token,
-                        line_index,
-                        line_offset,
-                        SemanticTokenType::Keyword,
-                        SemanticTokenModifiers::NONE,
-                        tokens,
-                    );
-                }
-
-                // Emit token for "on" keyword
-                if let Some(type_condition) = fragment.type_condition() {
-                    if let Some(on_token) = type_condition.on_token() {
-                        emit_token_for_syntax_token(
-                            &on_token,
-                            line_index,
-                            line_offset,
-                            SemanticTokenType::Keyword,
-                            SemanticTokenModifiers::NONE,
-                            tokens,
-                        );
-                    }
-                    // Emit token for type name
-                    if let Some(named_type) = type_condition.named_type() {
-                        if let Some(name) = named_type.name() {
-                            emit_token_for_syntax_node(
-                                name.syntax(),
-                                line_index,
-                                line_offset,
-                                SemanticTokenType::Type,
-                                SemanticTokenModifiers::NONE,
-                                tokens,
-                            );
-                        }
-                    }
-                }
-
-                // Get fragment's type condition for field deprecation checks
-                let type_name = fragment
-                    .type_condition()
-                    .and_then(|tc| tc.named_type())
-                    .and_then(|nt| nt.name())
-                    .map(|name| name.text().to_string());
-
-                // Process selection set
-                if let Some(selection_set) = fragment.selection_set() {
-                    collect_tokens_from_selection_set(
-                        &selection_set,
-                        type_name.as_deref(),
-                        schema_types,
-                        line_index,
-                        line_offset,
-                        tokens,
-                    );
-                }
-            }
-            _ => {
-                // Schema definitions (type, interface, etc.) - skip for now
-            }
-        }
-    }
-}
-
-/// Collect semantic tokens from a selection set
-#[allow(clippy::too_many_lines)]
-fn collect_tokens_from_selection_set(
-    selection_set: &apollo_parser::cst::SelectionSet,
-    parent_type_name: Option<&str>,
-    schema_types: Option<&std::collections::HashMap<Arc<str>, graphql_hir::TypeDef>>,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-    tokens: &mut Vec<SemanticToken>,
-) {
-    use apollo_parser::cst::{self, CstNode};
-
-    let parent_type = parent_type_name.and_then(|name| schema_types?.get(name));
-
-    for selection in selection_set.selections() {
-        match selection {
-            cst::Selection::Field(field) => {
-                if let Some(field_name_node) = field.name() {
-                    let field_name = field_name_node.text();
-
-                    // Check if field is deprecated
-                    let is_deprecated = parent_type
-                        .and_then(|pt| {
-                            pt.fields
-                                .iter()
-                                .find(|f| f.name.as_ref() == field_name.as_ref())
-                        })
-                        .is_some_and(|f| f.is_deprecated);
-
-                    let modifiers = if is_deprecated {
-                        SemanticTokenModifiers::DEPRECATED
-                    } else {
-                        SemanticTokenModifiers::NONE
-                    };
-
-                    emit_token_for_syntax_node(
-                        field_name_node.syntax(),
-                        line_index,
-                        line_offset,
-                        SemanticTokenType::Property,
-                        modifiers,
-                        tokens,
-                    );
-
-                    // Get field's return type for nested selection set
-                    let field_return_type = parent_type
-                        .and_then(|pt| {
-                            pt.fields
-                                .iter()
-                                .find(|f| f.name.as_ref() == field_name.as_ref())
-                        })
-                        .map(|f| f.type_ref.name.as_ref());
-
-                    // Recurse into nested selection set
-                    if let Some(nested_selection_set) = field.selection_set() {
-                        collect_tokens_from_selection_set(
-                            &nested_selection_set,
-                            field_return_type,
-                            schema_types,
-                            line_index,
-                            line_offset,
-                            tokens,
-                        );
-                    }
-                }
-            }
-            cst::Selection::FragmentSpread(spread) => {
-                // Emit token for "..." is not needed (punctuation)
-                // Emit token for fragment name
-                if let Some(name) = spread.fragment_name().and_then(|fn_| fn_.name()) {
-                    emit_token_for_syntax_node(
-                        name.syntax(),
-                        line_index,
-                        line_offset,
-                        SemanticTokenType::Function,
-                        SemanticTokenModifiers::NONE,
-                        tokens,
-                    );
-                }
-            }
-            cst::Selection::InlineFragment(inline) => {
-                // Emit token for "on" keyword and type name
-                if let Some(type_condition) = inline.type_condition() {
-                    if let Some(on_token) = type_condition.on_token() {
-                        emit_token_for_syntax_token(
-                            &on_token,
-                            line_index,
-                            line_offset,
-                            SemanticTokenType::Keyword,
-                            SemanticTokenModifiers::NONE,
-                            tokens,
-                        );
-                    }
-                    if let Some(named_type) = type_condition.named_type() {
-                        if let Some(name) = named_type.name() {
-                            emit_token_for_syntax_node(
-                                name.syntax(),
-                                line_index,
-                                line_offset,
-                                SemanticTokenType::Type,
-                                SemanticTokenModifiers::NONE,
-                                tokens,
-                            );
-                        }
-                    }
-                }
-
-                // Get type condition for nested selection set
-                let type_name = inline
-                    .type_condition()
-                    .and_then(|tc| tc.named_type())
-                    .and_then(|nt| nt.name())
-                    .map(|name| name.text().to_string());
-
-                let type_name_ref = type_name.as_deref().or(parent_type_name);
-
-                // Recurse into nested selection set
-                if let Some(selection_set) = inline.selection_set() {
-                    collect_tokens_from_selection_set(
-                        &selection_set,
-                        type_name_ref,
-                        schema_types,
-                        line_index,
-                        line_offset,
-                        tokens,
-                    );
-                }
-            }
-        }
-    }
-}
-
-/// Emit a semantic token for a syntax node
-#[allow(clippy::cast_possible_truncation)]
-fn emit_token_for_syntax_node(
-    node: &apollo_parser::SyntaxNode,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-    token_type: SemanticTokenType,
-    modifiers: SemanticTokenModifiers,
-    tokens: &mut Vec<SemanticToken>,
-) {
-    let offset: usize = node.text_range().start().into();
-    let len: u32 = node.text_range().len().into();
-
-    let (line, col) = line_index.line_col(offset);
-    tokens.push(SemanticToken::new(
-        Position::new(line as u32 + line_offset, col as u32),
-        len,
-        token_type,
-        modifiers,
-    ));
-}
-
-/// Emit a semantic token for a syntax token (keyword, punctuation, etc.)
-#[allow(clippy::cast_possible_truncation)]
-fn emit_token_for_syntax_token(
-    token: &apollo_parser::SyntaxToken,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-    token_type: SemanticTokenType,
-    modifiers: SemanticTokenModifiers,
-    tokens: &mut Vec<SemanticToken>,
-) {
-    let offset: usize = token.text_range().start().into();
-    let len: u32 = token.text_range().len().into();
-
-    let (line, col) = line_index.line_col(offset);
-    tokens.push(SemanticToken::new(
-        Position::new(line as u32 + line_offset, col as u32),
-        len,
-        token_type,
-        modifiers,
-    ));
-}
-
-/// Get field children for a type definition
-fn get_field_children(
-    structure: &graphql_hir::FileStructureData,
-    type_name: &str,
-    tree: &apollo_parser::SyntaxTree,
-    line_index: &graphql_syntax::LineIndex,
-    line_offset: u32,
-) -> Vec<DocumentSymbol> {
-    let Some(type_def) = structure
-        .type_defs
-        .iter()
-        .find(|t| t.name.as_ref() == type_name)
-    else {
-        return Vec::new();
-    };
-
-    let mut children = Vec::new();
-
-    for field in &type_def.fields {
-        if let Some(ranges) = find_field_definition_full_range(tree, type_name, &field.name) {
-            let range = adjust_range_for_line_offset(
-                offset_range_to_range(line_index, ranges.def_start, ranges.def_end),
-                line_offset,
-            );
-            let selection_range = adjust_range_for_line_offset(
-                offset_range_to_range(line_index, ranges.name_start, ranges.name_end),
-                line_offset,
-            );
-
-            let detail = format_type_ref(&field.type_ref);
-            children.push(
-                DocumentSymbol::new(
-                    field.name.to_string(),
-                    SymbolKind::Field,
-                    range,
-                    selection_range,
-                )
-                .with_detail(detail),
-            );
-        }
-    }
-
-    children
-}
+// Helper functions are now in feature modules
 
 /// Analyze selections recursively to calculate complexity
 #[allow(clippy::too_many_arguments)]
@@ -3180,7 +1886,7 @@ fn get_type_info(
 #[allow(clippy::needless_raw_string_hashes)]
 mod tests {
     use super::*;
-    use crate::helpers::{convert_position, convert_range, convert_severity};
+    use crate::helpers::{convert_position, convert_range, convert_severity, position_to_offset};
 
     #[test]
     fn test_analysis_host_creation() {
